src/main/kotlin/it/gasadvisor/gas_backend/repository/GasStationRepository.kt
package it.gasadvisor.gas_backend.repository

import it.gasadvisor.gas_backend.api.gas.station.contract.GetAllStationsResponse
import it.gasadvisor.gas_backend.model.GasStation
import org.springframework.data.jpa.repository.JpaRepository
import org.springframework.data.jpa.repository.Query
import org.springframework.stereotype.Repository

@Repository
interface GasStationRepository : JpaRepository<GasStation, Long> {

    @Query("select new it.gasadvisor.gas_backend.api.gas.station.contract.GetAllStationsResponse(g.id, g.latitude, g.longitude) " +
            "from GasStation g")
    fun findAllLocations(): List<GetAllStationsResponse>
}
